requirements/common.txt
Django>=3.2.16, <4
django-extensions==3.2.1
django-crispy-forms>=1.13.0, <=1.14.0
python-dateutil==2.8.2
Markdown>=3.4.1
Pygments==2.13.0
PyYAML==6.0
cradmin-legacy==4.1.2
psycopg2-binary==2.9.5

# For Dataporten login
django-allauth==0.51.0

#-e git://github.com/appressoas/django_cradmin.git#egg=django_cradmin

-e .
